完善了基本功能

+++ src/main/java/com/example/community/dto/QuestionDTO.java
@@ -5,15 +5,15 @@ import lombok.Data;
 
 @Data
 public class QuestionDTO {
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private Long gmtCreate;
     private Long gmtModified;
-    private Integer creator;
-    private Long commentCount;
-    private Long viewCount;
-    private Long likeCount;
+    private Long creator;
+    private Integer commentCount;
+    private Integer viewCount;
+    private Integer likeCount;
     private String tag;
     private User user;
 }
